packages/ml/requirements.txt
xgboost==2.1.3
scikit-learn==1.5.2
shap==0.46.0
pandas==2.2.3
numpy==1.26.4
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
